refactor(inference): extract Metal dispatch wrappers
Fixes #839
